Dagster updates

+++ src/assets/scrape_wiki.py
@@ -3,6 +3,9 @@ from bs4 import BeautifulSoup
 from datetime import date
 from src import aws
 from dagster import asset
+from src import utils
+from datetime import datetime, timedelta
+from pathlib import Path
 from src.assets.refresh_analytics import get_not_scraped_url
 dt = date.today().strftime("%Y-%m-%d")
 
@@ -25,19 +28,23 @@ def save_wiki_html(get_not_scraped_url):
     url = get_not_scraped_url['url']
 
     file_name = f"./html_files/{show}_{season}"
-    soup = scrapePage(url)
+    # Only scrape if it hasn't been scraped and
+    # the last scrape was at least 3 days ago.
+    if utils.does_file_exist(file_name):
 
-    text = soup.find_all('table', class_="wikitable sortable")
-    with open(file_name, "w") as f:
-        f.write(str(text))
-    print(f"Saved to {file_name}")
-    return {"html_file": file_name, "show": show, "season": season}
+        last_modified_time = utils.get_last_modified_time(file_name)
 
+        if datetime.now() - last_modified_time < timedelta(days=3):
+            print(f"{file_name} already exist, returning location.")
+            return {"html_file": file_name, "show": show, "season": season}
+    else:
+        soup = scrapePage(url)
 
-def saveToCsv(df, csv_name):
-    file = csv_name + '.csv'
-    df.to_csv(file, index=False)
-
+        text = soup.find_all('table', class_="wikitable sortable")
+        with open(file_name, "w") as f:
+            f.write(str(text))
+        print(f"Saved to {file_name}")
+        return {"html_file": file_name, "show": show, "season": season}
 
 def get_next_page_imdb(curr, soup):
     """
